requirements.txt
pyjwt==1.7.1
flask-jwt-extended==4.1.0
pytest

